Phase 3 (§8.3): derive write-mask width, !prof cap, type-driven range
- Write masks are u128 and the EMITTED width is i128 when the program has
  >64 state fields (dispatch.rs write_mask_type); the old idx < 64 gate
  silently dropped the 65th+ field.
- !prof branch-weight cap changed from 1000 to a power of two near
  i32::MAX / 2 (2^30): LLVM branch_weights are i32, and scaling keeps the
  sum <= cap so every weight fits i32 with no overflow and less ratio
  rounding. Iter bounds (<= 50M) now emit direct weights.
- type_driven_range derives [0, 2^(8*bytes)) from resolved.bytes instead of
  a hardcoded 1/2-byte table.

cargo test --lib: 1265 pass.

+++ src/backend/llvm/mod.rs
@@ -615,7 +615,10 @@ pub struct LlvmBackend {
     program_txns: Vec<String>,
     pub(crate) fused_to_first: HashMap<String, String>,
     pub(crate) sampled_triggers: HashMap<String, String>,
-    pub(crate) txn_write_masks: HashMap<String, u64>,
+    // 2026-07-31: Phase 3 (§8.3) — write masks are u128 so the 65th+ field is
+    // no longer silently dropped; the EMITTED width is i128 when the program
+    // has >64 state fields (write_mask_type), else i64.
+    pub(crate) txn_write_masks: HashMap<String, u128>,
     cell_thread_names: Vec<String>,
 
     // ── Reporting & Diagnostics ────────────────────────────
